apis: optimize extension APIs layout
